moved tessts into corresponding stages and updated test_coverage.md
